scope treaty font to warondisease
